pyproject.toml
[tool.poetry]
name = "scribe"
version = "0.1.0"
description = ""
authors = ["marcinplatek <maruplat@gmail.com>"]
readme = "README.md"
packages = [{include = "scribe", from = "src"}]

[tool.poetry.dependencies]
python = "~3.10"
openai-whisper = "^20230117"
streamlit = "^1.17.0"
pyperclip = "^1.8.2"


[tool.poetry.group.dev.dependencies]
watchdog = "^2.2.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
